refactor(db): extract migration stages into pipeline architecture
Convert monolithic `Migration` struct into a generic pipeline that
executes `MigrationStage` trait objects. Each stage (state updates,
receipt envelopes) is now independently defined in its own module.
