+++ crates/storage/db/src/models/envelope.rs
@@ -124,7 +124,7 @@ pub trait EnvelopePayload: Compress + Decompress + Debug + Clone + PartialEq + E
 ///
 /// Pre-envelope rows (raw postcard bytes without a magic prefix) are **not** handled by the
 /// envelope codec. They must be migrated to envelope format before they can be read.
-/// See [`Migration::backfill_receipt_envelopes`](crate::migration::Migration::backfill_receipt_envelopes).
+/// See [`ReceiptEnvelopeStage`](crate::migration::ReceiptEnvelopeStage).
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Envelope<T: EnvelopePayload> {
     pub inner: T,
